Move JSON serialization methods into OrleansJsonSerializer
